refactor: Controller Test 내 캘린더, 팀 캘린더 분리
